move factory to notifier package and switch remaining receivers to alerting
