+++ pkg/services/ngalert/api/tooling/definitions/provisioning_contactpoints.go
@@ -6,7 +6,7 @@ import (
 	"github.com/grafana/alerting/alerting/notifier/channels"
 
 	"github.com/grafana/grafana/pkg/components/simplejson"
-	ngchannels "github.com/grafana/grafana/pkg/services/ngalert/notifier/channels"
+	"github.com/grafana/grafana/pkg/services/ngalert/notifier"
 	"github.com/grafana/grafana/pkg/services/ngalert/notifier/channels_config"
 	"github.com/grafana/grafana/pkg/setting"
 )
@@ -107,7 +107,7 @@ func (e *EmbeddedContactPoint) Valid(decryptFunc channels.GetDecryptedValueFn) e
 	if e.Settings == nil {
 		return fmt.Errorf("settings should not be empty")
 	}
-	factory, exists := ngchannels.Factory(e.Type)
+	factory, exists := notifier.Factory(e.Type)
 	if !exists {
 		return fmt.Errorf("unknown type '%s'", e.Type)
 	}
