+++ pkg/services/ngalert/notifier/factory.go
@@ -1,4 +1,4 @@
-package channels
+package notifier
 
 import (
 	"strings"
@@ -7,18 +7,18 @@ import (
 )
 
 var receiverFactories = map[string]func(channels.FactoryConfig) (channels.NotificationChannel, error){
-	"prometheus-alertmanager": AlertmanagerFactory,
-	"dingding":                DingDingFactory,
-	"discord":                 DiscordFactory,
-	"email":                   EmailFactory,
-	"googlechat":              GoogleChatFactory,
-	"kafka":                   KafkaFactory,
-	"line":                    LineFactory,
+	"prometheus-alertmanager": channels.AlertmanagerFactory,
+	"dingding":                channels.DingDingFactory,
+	"discord":                 channels.DiscordFactory,
+	"email":                   channels.EmailFactory,
+	"googlechat":              channels.GoogleChatFactory,
+	"kafka":                   channels.KafkaFactory,
+	"line":                    channels.LineFactory,
 	"opsgenie":                channels.OpsgenieFactory,
 	"pagerduty":               channels.PagerdutyFactory,
 	"pushover":                channels.PushoverFactory,
 	"sensugo":                 channels.SensuGoFactory,
-	"slack":                   SlackFactory,
+	"slack":                   channels.SlackFactory,
 	"teams":                   channels.TeamsFactory,
 	"telegram":                channels.TelegramFactory,
 	"threema":                 channels.ThreemaFactory,
